requirements.txt
flask>=3.0
anthropic>=0.40
twilio>=9.0
python-dotenv>=1.0
